Add attribution for clever line length workaround
Found an interesting way to show an error when line lengths are too long on
StackOverflow for version of Vim that don't support it natively.  I added an
attribution to the author of that StackOverflow comment because I like the thing
a lot and that's the right thing to do.