Add Syntastic settings to quiet HTML errors
Syntastic (which I wasn't even aware was installed, must be from the Thoughtbot
dotfiles, which would make sense) is processing my Handlebars templates as HTML
files and throwing all sorts of errors.  Thanks to this Github Issue:

https://github.com/mustache/vim-mustache-handlebars/issues/6

I found out how to silence the errors, which isn't perfect but is better than
nothing.