Check for solarized theme before applying it
If you try to open vim without the plugins installed (including when you actually install the plugins the first time) you would get an error saying that you're missing a the theme.  Now, we look for the theme first and only use it if it exists, meaning we no longer get that warning.  If the entire plugin set has been successfully installed, then it should find the theme without issue.  It looks for the theme in the location that Vundle installs it to