Dynamically enable JS makers
After getting really tired of setting the JS makers for different projects, I decided to just gynamically check for them.  For each configuration file specified, it'll walk up the directory tree and check if the file exists, enabling the corresponding maker only if we can find a configuration file for it.